requirements.txt
python>=3.8
paddle==2.2.1
pyyaml
easydict
tqdm